requirements.txt
# Core dependencies
Flask==2.3.3
Werkzeug==3.0.1
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3

# Database and ORM
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5

# Authentication and Forms
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email_validator==2.1.0

# Configuration and Environment
python-dotenv==1.0.0
pytz==2023.3

# Security
pyopenssl==23.3.0
cryptography==41.0.5

# AI Content Detection Dependencies
statistics==1.0.3.5
nltk==3.8.1
requests==2.31.0

# Document Import Dependencies
python-docx==0.8.11
PyPDF2==3.0.1
openai==0.28.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
psutil==5.9.6
python-dateutil==2.8.2
pandas==2.0.3